Support slicing sub image (#21)
Add argument --parentzoomlevel and --indexforzoom. If a valid parentzoomlevel value is provided, zoomrange and targetrange values will be ignored, it will slice all the tiles for the current zoomlevel value. For example, for the four level 5 sub images from one level 5 image, --zoomlevel should be 5, --parentzoomlevel should be 6, and --indexforzoom should be a number in 0-3. It will slice 1024 tiles for each input image and give them appropriate naming. Running all four of them should output the same results as slicing all the tiles for one level 6 input image.
Replace startzoomrangetoslice and endzoomrangetoslice to be zoomrangetoslice, starttargetrange and endtargetrange to be targetrangetoslice in config to clean up.